test: 로그인 요청 테스트 코드 추가
