+++ src/test/java/com/moabam/api/application/AuthenticationServiceTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 
-import java.io.IOException;
 import java.util.List;
 
 import org.assertj.core.api.Assertions;
@@ -16,12 +15,11 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
-import org.mockito.Mockito;
+import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatusCode;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.util.ReflectionTestUtils;
@@ -31,18 +29,19 @@ import com.moabam.api.dto.AuthorizationCodeRequest;
 import com.moabam.api.dto.AuthorizationCodeResponse;
 import com.moabam.api.dto.AuthorizationTokenResponse;
 import com.moabam.api.dto.OAuthMapper;
-import com.moabam.global.common.util.GlobalConstant;
 import com.moabam.global.config.OAuthConfig;
 import com.moabam.global.error.exception.BadRequestException;
 import com.moabam.global.error.model.ErrorMessage;
-
-import jakarta.servlet.http.HttpServletResponse;
 
 @ExtendWith(MockitoExtension.class)
 class AuthenticationServiceTest {
 
 	@InjectMocks
 	AuthenticationService authenticationService;
+
+	@Mock
+	OAuth2AuthorizationServerRequestService oAuth2AuthorizationServerRequestService;
+
 	OAuthConfig oauthConfig;
 	AuthenticationService noPropertyService;
 	OAuthConfig noOAuthConfig;
@@ -60,7 +59,7 @@ class AuthenticationServiceTest {
 			new OAuthConfig.Provider(null, null, null),
 			new OAuthConfig.Client(null, null, null, null, null)
 		);
-		noPropertyService = new AuthenticationService(noOAuthConfig);
+		noPropertyService = new AuthenticationService(noOAuthConfig, oAuth2AuthorizationServerRequestService);
 
 	}
 
@@ -86,43 +85,17 @@ class AuthenticationServiceTest {
 		);
 	}
 
-	@DisplayName("인가코드 URI 생성 성공")
+	@DisplayName("redirect 로그인페이지 성공")
 	@Test
-	void authorization_code_uri_generate_success() throws IOException {
+	void redirect_loginPage_success() {
 		// given
-		String uri = "https://authorization/url?"
-			+ "response_type=code&"
-			+ "client_id=testtestetsttest&"
-			+ "redirect_uri=http://redirect/url&scope=profile_nickname,profile_image";
-
-		MockHttpServletResponse mockHttpServletResponse = mockHttpServletResponse = new MockHttpServletResponse();
+		MockHttpServletResponse mockHttpServletResponse = new MockHttpServletResponse();
 
 		// when
 		authenticationService.redirectToLoginPage(mockHttpServletResponse);
 
 		// then
-		assertThat(mockHttpServletResponse.getContentType())
-			.isEqualTo(MediaType.APPLICATION_FORM_URLENCODED + GlobalConstant.CHARSET_UTF_8);
-		assertThat(mockHttpServletResponse.getRedirectedUrl()).isEqualTo(uri);
-	}
-
-	@DisplayName("redirect 실패 테스트")
-	@Test
-	void redirect_fail_test() {
-		// given
-		HttpServletResponse mockHttpServletResponse = Mockito.mock(HttpServletResponse.class);
-
-		try {
-			doThrow(IOException.class).when(mockHttpServletResponse).sendRedirect(any(String.class));
-
-			assertThatThrownBy(() -> {
-				// When + Then
-				authenticationService.redirectToLoginPage(mockHttpServletResponse);
-			}).isExactlyInstanceOf(BadRequestException.class)
-				.hasMessage(ErrorMessage.REQUEST_FAILED.getMessage());
-		} catch (Exception ignored) {
-
-		}
+		verify(oAuth2AuthorizationServerRequestService).loginRequest(eq(mockHttpServletResponse), anyString());
 	}
 
 	@DisplayName("인가코드 반환 실패 테스트")
